refactor Create Person classes into CreateUser class
