+++ Project/app.rb
@@ -1,35 +1,7 @@
-require_relative 'student'
-require_relative 'teacher'
+require_relative 'create_user'
 require_relative 'book'
 require_relative 'rental'
 
-class CreateStudent
-  def create_student(people)
-    print 'Age: '
-    age = gets.chomp.to_i
-    print 'Name: '
-    name = gets.chomp
-    print 'Has parent permission? (Y/N): '
-    permission = gets.chomp.downcase == 'y'
-    people << if permission
-                 Student.new(age, name)
-               else
-                 Student.new(age, name, parent_permission: false)
-               end
-  end
-end
-
-class CreateTeacher
-  def create_teacher(people)
-    print 'Age: '
-    age = gets.chomp.to_i
-    print 'Name: '
-    name = gets.chomp
-    print 'Specialization: '
-    specialization = gets.chomp
-    people << Teacher.new(age, specialization, name: name)
-  end
-end
 
 class Create_book
   def create_book(books)
@@ -43,7 +15,6 @@ class Create_book
     puts '-------------------------------'
   end
 end
-
 
 class App
   def initialize
@@ -72,15 +43,14 @@ class App
     puts '---------------------------------'
   end
 
-  
   def create_person
     puts 'Do you want to create a student (1) or a teacher (2)?'
     person_type = gets.chomp.to_i
     case person_type
     when 1
-      CreateStudent.new.create_student(@people)
+      CreateUser.new.create_student(@people)
     when 2
-      CreateTeacher.new.create_teacher(@people)
+      CreateUser.new.create_teacher(@people)
     else
       puts 'Person type not recognized'
     end
